Extract shared profiling helpers to reduce JS duplication
Move base64 decode/download logic from job-detail.js and task-detail.js
into shared/profiling.js with a reusable profileAndDownload() function.
