+++ lib/iris/src/iris/cluster/static/controller/job-detail.js
@@ -8,6 +8,7 @@ import htm from 'htm';
 import { controllerRpc } from '/static/shared/rpc.js';
 import { formatBytes, formatDuration, formatTimestamp, stateToName, timestampFromProto } from '/static/shared/utils.js';
 import { InfoRow, InfoCard } from '/static/shared/components.js';
+import { profileAndDownload } from '/static/shared/profiling.js';
 
 const html = htm.bind(h);
 
@@ -191,34 +192,7 @@ function JobDetailApp() {
   async function handleProfileTask(taskId) {
     setProfilingTaskId(taskId);
     try {
-      const resp = await controllerRpc('ProfileTask', {
-        taskId,
-        durationSeconds: 10,
-        format: 'speedscope',
-      });
-      if (resp.error) {
-        alert('Profile error: ' + resp.error);
-        return;
-      }
-      // profileData is base64-encoded bytes in JSON
-      const raw = atob(resp.profileData);
-      const bytes = new Uint8Array(raw.length);
-      for (let i = 0; i < raw.length; i++) bytes[i] = raw.charCodeAt(i);
-      const blob = new Blob([bytes], { type: 'application/json' });
-      const url = URL.createObjectURL(blob);
-
-      // Open speedscope.app in a new tab with the profile data
-      const speedscopeUrl = 'https://www.speedscope.app/#profileURL=' + encodeURIComponent(url);
-      window.open(speedscopeUrl, '_blank');
-
-      // Also download the JSON file for manual viewing
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = 'profile-' + taskId.replace(/\//g, '_') + '.json';
-      a.click();
-
-      // Clean up the object URL after a delay to ensure download completes
-      setTimeout(() => URL.revokeObjectURL(url), 1000);
+      await profileAndDownload(controllerRpc, taskId);
     } catch (e) {
       alert('Profile failed: ' + e.message);
     } finally {
