+++ lib/iris/src/iris/cluster/static/worker/task-detail.js
@@ -3,6 +3,7 @@ import { useState, useEffect } from 'preact/hooks';
 import htm from 'htm';
 import { stateToName, formatTimestamp, timestampFromProto } from '/static/shared/utils.js';
 import { workerRpc } from '/static/shared/rpc.js';
+import { profileAndDownload } from '/static/shared/profiling.js';
 
 const html = htm.bind(h);
 
@@ -105,7 +106,7 @@ function ProfileSection({ task, onProfile, profiling }) {
     >
       ${profiling ? 'Profiling (10s)...' : 'Profile with py-spy'}
     </button>
-    <p style="color:#57606a;font-size:12px;margin-top:8px">Captures a 10-second profile in speedscope format. Opens in speedscope.app and downloads JSON.</p>
+    <p style="color:#57606a;font-size:12px;margin-top:8px">Captures a 10-second profile. Open the downloaded file in <a href="https://www.speedscope.app" target="_blank">speedscope.app</a>.</p>
   </div>`;
 }
 
@@ -128,33 +129,7 @@ function TaskDetail() {
   async function handleProfile() {
     setProfiling(true);
     try {
-      const resp = await workerRpc('ProfileTask', {
-        taskId,
-        durationSeconds: 10,
-        format: 'speedscope',
-      });
-      if (resp.error) {
-        alert('Profile error: ' + resp.error);
-        return;
-      }
-      const raw = atob(resp.profileData);
-      const bytes = new Uint8Array(raw.length);
-      for (let i = 0; i < raw.length; i++) bytes[i] = raw.charCodeAt(i);
-      const blob = new Blob([bytes], { type: 'application/json' });
-      const url = URL.createObjectURL(blob);
-
-      // Open speedscope.app in a new tab with the profile data
-      const speedscopeUrl = 'https://www.speedscope.app/#profileURL=' + encodeURIComponent(url);
-      window.open(speedscopeUrl, '_blank');
-
-      // Also download the JSON file for manual viewing
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = 'profile-' + taskId.replace(/\//g, '_') + '.json';
-      a.click();
-
-      // Clean up the object URL after a delay to ensure download completes
-      setTimeout(() => URL.revokeObjectURL(url), 1000);
+      await profileAndDownload(workerRpc, taskId);
     } catch (e) {
       alert('Profile failed: ' + e.message);
     } finally {
